// ts/Model/Discord/ApplicationCommandsResult.ts
export type ApplicationCommandsResult = {
  id: string;
  description_localizations: unknown;
  application_id: string;
  version: string;
  dm_permission: boolean;
  default_permission: boolean;
  default_member_permissions: unknown;
  nsfw: boolean;
  type: number;
  name: string;
  name_localizations: unknown;
  description: string;
};